# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NGO HomeSuite API
  version: 0.1.0-beta
  description: |
    Starter OpenAPI contract for internal/beta integrations.
    This contract emphasizes stable operationIds and reusable schemas to support
    downstream tooling (client generation, docs, and contract checks).
servers:
  - url: /
paths:
...
  /api/v1/workflows:
    get:
      operationId: getV1WorkflowDefinitions
      summary: List deterministic workflow definitions in V2 runtime
      tags: [WorkflowsV1]
      responses:
        '200':
          description: Workflow definition list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  workflow_types:
                    type: array
                    items:
                      type: string
                required: [ok, workflow_types]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workflows/instances:
    post:
      operationId: postV1CreateWorkflowInstance
      summary: Create a workflow instance
      tags: [WorkflowsV1]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowInstanceRequest'
      responses:
        '200':
          description: Created workflow instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1WorkflowInstanceEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/workflows/instances/{instanceId}/events:
    post:
      operationId: postV1ApplyWorkflowEvent
      summary: Apply event transition to workflow instance
      tags: [WorkflowsV1]
      parameters:
        - in: path
          name: instanceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyWorkflowEventRequest'
      responses:
        '200':
          description: Updated workflow instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1WorkflowInstanceEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /ai/minion/chat:
    post:
      operationId: postMinionChat
      summary: Query HomeSuite Minion
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MinionChatRequest'
      responses:
        '200':
          description: Minion response envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MinionChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MinionActionApprovalItem:
      type: object
      properties:
        tool:
          type: string
        token:
          type: string
      required: [tool]
    MinionChatRequest:
      type: object
      properties:
        prompt:
          type: string
        context:
          type: object
          additionalProperties: true
        allow_actions:
          type: boolean
        approved_actions:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/MinionActionApprovalItem'
      required: [prompt]
    SourceSnippet:
      type: object
      additionalProperties: true
    ToolAction:
      type: object
      additionalProperties: true
...
    CreateWorkflowInstanceRequest:
      type: object
      properties:
        org_id:
          type: string
        workflow_type:
          type: string
      required: [org_id, workflow_type]
    ApplyWorkflowEventRequest:
      type: object
      properties:
        org_id:
          type: string
        event_type:
          type: string
        actor_id:
          type: string
        role:
          type: string
        payload:
          type: object
          additionalProperties: true
      required: [org_id, event_type, actor_id, role]
    V1WorkflowInstance:
      type: object
      properties:
        instance_id:
          type: string
        org_id:
          type: string
        workflow_type:
          type: string
        current_step:
          type: string
        status:
          type: string
        history:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [instance_id, org_id, workflow_type, current_step, status]
    V1WorkflowInstanceEnvelope:
      type: object
      properties:
        ok:
          type: boolean
        instance:
          $ref: '#/components/schemas/V1WorkflowInstance'
      required: [ok, instance]
    MinionChatResponse:
      type: object
      properties:
        response:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceSnippet'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ToolAction'
        redactions:
          type: integer
        mode:
          type: string
      required: [response, sources, actions, mode]
...
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: Service unavailable or disabled
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: session
security:
  - sessionCookieAuth: []